Change executor type for test step in circleci config.
Executor type was changed to machine in order to work correctly with docker containter created directly from code.
Main problem is that setup_remote_docker command is running VM with docker env.
And it's impossible to connect to this VM.
Maybe there is some workaround or maybe it's better just to create your own custom docker image with golang and docker inside.
But for now solution is just simply to change executor type.
Also add step to update go version to 1.12.6, cause actual version in provided image is 1.10.3 and it doesn't support modules.