Add Join- and FinishTournament transactional methods to storage pkg.
These methods contain logic that covers more then 1 operation of write in or read from mongodb.
Need to notice that transactions in mongodb work only with replica set or shards.
This little problem is easy to solve following official documentation, but still need to be remembered.
Add JoinTournament and FinishTournament methods to Service interface and to storage_mock.go file.