Change logic of finishing tournament.
The main problem was that when we finish tournament we don't know winner ID.
So now we pass winner user ID everywhere to close this logic hole.
So in server pkg we add winnerUserID struct and now we check request body in finishTournament endpoint.
Also finishTournament endpoint tests were rewritten to follow new approach.
In storage pkg we add winnerUserID variable to FinishTournament func.
Also Service interface was updated here. As a result storage_mock.go file was edited too.